+++ shared/src/iosMain/kotlin/eu/heha/conifer/IosConiferApp.kt
@@ -8,6 +8,7 @@ object IosConiferApp {
             isDebug = isDebug,
             platform = IosPlatform,
             databaseInitializer = IosDatabaseInitializer,
+            syncPrefsInitializer = IosSyncPrefsInitializer,
             clipboardController = IosClipboardController
         )
     }